+++ cpp/PInvokeNativeHelper/include/PInvokeNativeHelper/StreamHandle.h
@@ -51,8 +51,8 @@ extern "C"
 	int64_t Stream_Length(StreamHandle *handle);
 	void Stream_SetLength(StreamHandle *handle, int64_t value);
 
-	int64_t Stream_Read(StreamHandle *handle, uint8_t *dst_buf, int64_t offset, int64_t count);
-	void Stream_Write(StreamHandle *handle, uint8_t *src_buf, int64_t offset, int64_t count);
+	int32_t Stream_Read(StreamHandle *handle, uint8_t *dst_buf, int32_t offset, int32_t count);
+	void Stream_Write(StreamHandle *handle, uint8_t *src_buf, int32_t offset, int32_t count);
 
 	void Stream_Flush(StreamHandle *handle);
 	void Stream_Close(StreamHandle *handle);
