+++ cpp/jccpp/include/jccpp/stream/CircleBufferMemoryStream.h
@@ -13,23 +13,23 @@ namespace jc
 	class CircleBufferMemoryStream :public Stream
 	{
 	public:
-		CircleBufferMemoryStream(int64_t size);
+		CircleBufferMemoryStream(int32_t size);
 		~CircleBufferMemoryStream();
 
 	private:
 		uint8_t *_buffer;
-		int64_t _buffer_size;
+		int32_t _buffer_size;
 
 		/// <summary>
 		///		头部。读取的时候递增头部。头部指向的位置是有效元素。（前提是缓冲区不为空）
 		/// </summary>
-		int64_t _head;
+		int32_t _head;
 
 		/// <summary>
 		///		尾部。写入的时候递增尾部。尾部指向的位置是无效元素。
 		///		尾部 - 1 才是指向有效元素。（前提是缓冲区不为空）
 		/// </summary>
-		int64_t _tail;
+		int32_t _tail;
 
 		/// <summary>
 		///		队列是否已满。当头指针和尾指针重叠时，有 2 种情况：
@@ -39,17 +39,17 @@ namespace jc
 		/// </summary>
 		bool _is_full;
 
-		void AddHead(int64_t value);
-		void AddTail(int64_t value);
-		int64_t ReadNonCircular(uint8_t *dst_buf, int64_t offset, int64_t count);
-		void WriteNonCircular(uint8_t *src_buf, int64_t offset, int64_t count);
+		void AddHead(int32_t value);
+		void AddTail(int32_t value);
+		int32_t ReadNonCircular(uint8_t *dst_buf, int32_t offset, int32_t count);
+		void WriteNonCircular(uint8_t *src_buf, int32_t offset, int32_t count);
 
 	public:
 		/// <summary>
 		///		返回内部缓冲区的大小，也是此流所能达到的最大长度。
 		/// </summary>
 		/// <returns></returns>
-		int64_t MaxSize() const
+		int32_t MaxSize() const
 		{
 			return _buffer_size;
 		}
@@ -67,7 +67,7 @@ namespace jc
 		///		本流内部的缓冲区剩余空间。单位：字节。
 		/// </summary>
 		/// <returns></returns>
-		int64_t BufferAvailableSpace() const;
+		int32_t BufferAvailableSpace() const;
 
 		/// <summary>
 		///		清空流
@@ -91,8 +91,8 @@ namespace jc
 		/// <param name="value"></param>
 		void SetLength(int64_t value) override;
 
-		int64_t Read(uint8_t *dst_buf, int64_t offset, int64_t count) override;
-		void Write(uint8_t *src_buf, int64_t offset, int64_t count) override;
+		int32_t Read(uint8_t *dst_buf, int32_t offset, int32_t count) override;
+		void Write(uint8_t *src_buf, int32_t offset, int32_t count) override;
 
 		/// <summary>
 		///		调用后什么都不会发生。这是一个空的函数。
